Reworked the code from single file to folder structure.
