+++ webviz_subsurface/plugins/_group_tree/__init__.py
@@ -1,0 +1,1 @@
+from .group_tree import GroupTree
