+++ webviz_subsurface/plugins/_group_tree/controllers/__init__.py
@@ -1,0 +1,1 @@
+from .controllers import controllers
