+++ webviz_subsurface/plugins/_group_tree/views/__init__.py
@@ -1,0 +1,1 @@
+from .main_view import main_view
